feat(email-cascade): reorder providers, dynamic Resend cap, engagement-sorted job alerts
Resend was capped near its billing-cycle limit while sitting second in the
provider cascade, and job-alert sends went out in arbitrary order — meaning a
day that hit the shared daily send cap could skip the subscribers most likely
to open/click. This reworks the cascade so Resend absorbs load last and paces
itself to land under 50k/cycle, and sorts job-alert recipients by engagement
so a capped-out day still reaches the audience most likely to convert.

- Reorder scripts/lib/email-cascade.mjs PROVIDERS so Resend is no longer
  second in line; other providers absorb volume first.
- Add computeResendDynamicDailyLimit/fetchResendCycleUsage/resendCycleBounds
  to pace Resend sends against its cycle-anchored (6th-of-month) 50k quota
  instead of a static per-day number.
- Sort scripts/send-job-alerts.mjs recipients by click/open recency so a
  capped-out day still favors the most-engaged subscribers.
- Route the remaining direct-Resend-bypass call sites (calculator report,
  publisher renewal/application confirmations, Stripe publisher
  confirmation, publisher ad blasts, journalist "article is live"
  notification, GSC indexation monitor alert) through the shared cascade so
  they benefit from the same pacing/fallback instead of hammering Resend
  directly.
- Move the cascade implementation to functions/src/emailCascade.js (the
  no-bundler-safe location Firebase Functions can import) and shrink
  scripts/lib/email-cascade.mjs to a thin re-export shim.
- Thread an optional AbortSignal (opts.signal) through
  sendEmailCascade -> sendSingleThrottled -> sendSingle -> each provider's
  fetch(), so callers that used to bound a raw fetch() (e.g.
  notify-journalist-article-live.mjs) can still bound the send now that it
  goes through the cascade. Backward compatible: no existing caller passes
  signal, so it defaults to undefined everywhere else.
- Add scripts/check-email-quotas.mjs + .github/workflows/check-email-quotas.yml
  to alert when Resend burn-rate risks the monthly cap before renewal.
- Update tests/functions/error-response.test.ts to mock
  functions/src/emailCascade.js's isProviderConfigured/sendEmailCascade
  directly (matching the handler's actual cascade-routed signature) instead
  of an injected resendClient/resendApiKey that the handler never accepted.

Verified: 476/476 tests green across the 27 affected/adjacent test files;
GitNexus detect_changes reports low risk with zero affected execution
processes across all 21 changed symbols.
