+++ scripts/check-email-quotas.mjs
@@ -9,12 +9,22 @@
  * reports month-to-date consumption against the 3000/mo allowance via the
  * GraphQL Analytics API (emailSendingAdaptiveGroups).
  *
+ * It also watches Resend's billing-cycle burn rate (cycle anchored on the 6th
+ * of each month, same as the cascade's own dynamic daily cap) and opens a
+ * GitHub issue if usage is running ahead of the pace needed to land at/under
+ * the 50k/mo quota by renewal — a signal the dynamic cap alone can't surface,
+ * since it only throttles FUTURE sends, it doesn't tell a human the burn rate
+ * looks wrong (e.g. verification blind spot, cap bypassed, real demand spike).
+ *
  * Usage (env must already carry the provider keys — run load-rc-env first):
  *   eval "$(node scripts/load-rc-env.mjs)" && node scripts/check-email-quotas.mjs
+ *   node scripts/check-email-quotas.mjs --dry-run   # print, never open an issue
  *
  * Exit code is always 0 — this is a read-only diagnostic, never a gate.
  */
 
+import { execFileSync } from 'node:child_process';
+import { pathToFileURL } from 'node:url';
 import {
   PROVIDERS,
   syncQuotasFromAPIs,
@@ -22,7 +32,19 @@ import {
   isProviderConfigured,
   fetchCloudflareUsage,
   fetchCloudflareDeliveryStats,
+  fetchResendCycleUsage,
 } from './lib/email-cascade.mjs';
+
+// 15% slack over the expected pace before alerting — daily sends are lumpy
+// (newsletter days vs quiet days), so a small overshoot is normal noise, not
+// a real risk to the 50k/mo cap. Below 20% of the monthly quota consumed we
+// skip the check entirely: early in the cycle a handful of transactional
+// sends can look like a huge overshoot in ratio terms while being a
+// negligible absolute risk (e.g. 3 sent vs 1 expected = "300% of pace").
+const RESEND_PACE_BUFFER = 1.15;
+const RESEND_PACE_MIN_USAGE_RATIO = 0.2;
+const RESEND_PACING_ISSUE_TITLE = 'Resend: burn rate ahead of pace for 50k/mo renewal';
+const WORKFLOW_NAME = 'Email Quota Check';
 
 function todayUTC() {
   return new Date().toISOString().slice(0, 10);
@@ -32,7 +54,102 @@ function firstOfMonthUTC() {
   return todayUTC().slice(0, 7) + '-01';
 }
 
-async function main() {
+function pct(n) {
+  return `${(n * 100).toFixed(1)}%`;
+}
+
+/**
+ * Pure ratio math, split out from checkResendPacing() so it's testable
+ * without mocking the network (mirrors dmarc-monitor.mjs's analyze()).
+ * Returns a signal object when usage is meaningfully ahead of pace, or null
+ * when on-pace / too-early-to-tell (fail-safe: never alert on noise).
+ */
+export function computePacingSignal({ count, monthlyLimit, cycleStart, cycleEnd, now }) {
+  const totalCycleMs = cycleEnd.getTime() - cycleStart.getTime();
+  const elapsedMs = now - cycleStart.getTime();
+  const expectedRatio = Math.min(1, Math.max(0, elapsedMs / totalCycleMs));
+  const actualRatio = count / monthlyLimit;
+  const daysRemaining = Math.max(1, Math.ceil((cycleEnd.getTime() - now) / 86_400_000));
+
+  const signal = { count, monthlyLimit, actualRatio, expectedRatio, daysRemaining, cycleStart, cycleEnd };
+
+  if (actualRatio < RESEND_PACE_MIN_USAGE_RATIO) return { ...signal, aheadOfPace: false };
+  if (actualRatio <= expectedRatio * RESEND_PACE_BUFFER) return { ...signal, aheadOfPace: false };
+  return { ...signal, aheadOfPace: true };
+}
+
+/**
+ * Fetches Resend's cycle-to-date usage and runs it through computePacingSignal().
+ * Returns null when unconfigured/unverifiable (fail-safe) or on-pace.
+ */
+async function checkResendPacing() {
+  if (!isProviderConfigured('resend')) return null;
+  const provider = PROVIDERS.find((p) => p.id === 'resend');
+  const { count, truncated, cycleStart, cycleEnd } = await fetchResendCycleUsage();
+  if (truncated) {
+    console.log('\n⚠️  Resend cycle usage unverifiable (paging truncated) — cannot assess pacing.');
+    return null;
+  }
+
+  const signal = computePacingSignal({ count, monthlyLimit: provider.monthlyLimit, cycleStart, cycleEnd, now: Date.now() });
+  console.log(`\n📈 Resend cycle pacing: used ${signal.count}/${signal.monthlyLimit} (${pct(signal.actualRatio)}), expected pace ${pct(signal.expectedRatio)}, ${signal.daysRemaining}d remaining until renewal (${cycleEnd.toISOString().slice(0, 10)}).`);
+
+  return signal.aheadOfPace ? signal : null;
+}
+
+export function buildPacingIssueBody(signal) {
+  return [
+    '## 📈 Resend: burn rate ahead of pace',
+    '',
+    `Ciclo di fatturazione **${signal.cycleStart.toISOString().slice(0, 10)} → ${signal.cycleEnd.toISOString().slice(0, 10)}** (rinnovo il 6 del mese).`,
+    '',
+    `Usati **${signal.count} / ${signal.monthlyLimit}** (**${pct(signal.actualRatio)}**) contro un ritmo atteso del **${pct(signal.expectedRatio)}**.`,
+    `Giorni rimanenti al rinnovo: **${signal.daysRemaining}**.`,
+    '',
+    'Il cap dinamico su Resend (`computeResendDynamicDailyLimit` in `functions/src/emailCascade.js`)',
+    'rallenta automaticamente gli invii FUTURI in base a questo stesso calcolo, ma non segnala a un',
+    'umano quando il ritmo attuale è già fuori scala — questo alert copre quel buco (cap bypassato,',
+    'lookup usage inaffidabile ripetuto, picco reale di domanda).',
+    '',
+    '### Possibili cause',
+    '- Un invio ha bypassato la cascade (bypass diretto Resend non ancora individuato).',
+    '- Il riordino dei provider non sta assorbendo abbastanza volume prima di Resend.',
+    '- Picco reale di domanda (newsletter/job-alert più grandi del solito).',
+    '',
+    '_Aperta automaticamente dal workflow Email Quota Check. Chiudi quando il ritmo rientra sotto pace._',
+  ].join('\n');
+}
+
+function createIssue({ title, description, priority }) {
+  // Mirrors dmarc-monitor.mjs's createIssue(). `automation` label keeps the
+  // deterministic issue-triage classifier from routing this to an auto-fixer.
+  execFileSync(
+    'node',
+    [
+      'scripts/lib/github-issue-creator.mjs',
+      '--title', title,
+      '--description', description,
+      '--priority', String(priority),
+      '--label', 'automation',
+      '--workflow', WORKFLOW_NAME,
+    ],
+    { stdio: 'inherit' },
+  );
+}
+
+function parseArgs(argv) {
+  const opts = { dryRun: false };
+  for (const a of argv.slice(2)) {
+    if (a === '--dry-run') opts.dryRun = true;
+    else {
+      console.error(`Unknown argument: ${a}`);
+      process.exit(2);
+    }
+  }
+  return opts;
+}
+
+async function main(opts) {
   console.log('🔎 Email provider quota check\n');
 
   const configured = PROVIDERS.filter(p => isProviderConfigured(p.id)).map(p => p.id);
@@ -77,10 +194,27 @@ async function main() {
       }
     }
   }
+
+  const pacingSignal = await checkResendPacing();
+  if (pacingSignal) {
+    if (opts.dryRun) {
+      console.log(`\n[dry-run] Would open issue "${RESEND_PACING_ISSUE_TITLE}" (priority 2).`);
+    } else {
+      console.log(`\n🚨 Resend ahead of pace — opening/updating GitHub issue.`);
+      createIssue({
+        title: RESEND_PACING_ISSUE_TITLE,
+        description: buildPacingIssueBody(pacingSignal),
+        priority: 2,
+      });
+    }
+  }
 }
 
-main().catch(err => {
-  console.error('check-email-quotas failed:', err?.message || err);
-  // Diagnostic only — never fail a pipeline on this.
-  process.exit(0);
-});
+if (import.meta.url === pathToFileURL(process.argv[1]).href) {
+  const opts = parseArgs(process.argv);
+  main(opts).catch(err => {
+    console.error('check-email-quotas failed:', err?.message || err);
+    // Diagnostic only — never fail a pipeline on this.
+    process.exit(0);
+  });
+}
